refactured dialoges
