+++ src/components/BatchEdit.vue
@@ -1,7 +1,7 @@
 <template>
   <v-dialog v-model="dialog" max-width="600px">
     <template v-slot:activator="{ on }">
-      <v-btn color="primary" dark @click="$refs.dialog.$emit('openDialog')" v-on="on">Batch Edit</v-btn>
+      <v-btn color="primary" dark @click="$refs.dialog.$emit('openBatchEditDialog')" v-on="on">Export Board</v-btn>
     </template>
     <batch-edit-dialog ref="dialog" @close="dialog = false" />
   </v-dialog>
